+++ DesafioB3.Server/Features/CalculateAmount/CalculateAmountRequest.cs
@@ -1,4 +1,4 @@
 namespace DesafioB3.API.Features.CalculateAmount
 {
-    public record CalculateAmountRequest(decimal InitialValue, decimal Quantity);
+    public record CalculateAmountRequest(decimal InitialValue, int Quantity);
 }
